+++ examples/misc/server-rendering/src/index.js
@@ -2,13 +2,19 @@ import React from "react";
 import ReactDOM from "react-dom";
 import Browser from "@hickory/browser";
 import { curi } from "@curi/router";
-import { CuriProvider } from "@curi/react-dom";
+import { curiProvider } from "@curi/react-dom";
 
 import routes from "./routes";
-import { responseHandler } from "./render";
+import { renderResponse } from "./render";
 
 const history = Browser();
 const router = curi(history, routes);
+const Router = curiProvider(router);
 const root = document.getElementById("root");
 
-router.respond(responseHandler);
+router.respond(() => {
+  ReactDOM.hydrate(
+    <Router>{renderResponse}</Router>,
+    document.getElementById("root")
+  );
+});