+++ examples/misc/side-effect/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { curi } from "@curi/router";
 import Browser from "@hickory/browser";
 import createTitleSideEffect from "@curi/side-effect-title";
-import { CuriProvider } from "@curi/react-dom";
+import { curiProvider } from "@curi/react-dom";
 
 import routes from "./routes";
 import NavLinks from "./components/NavLinks";
@@ -15,9 +15,10 @@ const history = Browser();
 const router = curi(history, routes, {
   sideEffects: [{ effect: setTitle }]
 });
+const Router = curiProvider(router);
 
 ReactDOM.render(
-  <CuriProvider router={router}>
+  <Router>
     {({ response }) => {
       const { params, body: Body } = response;
       return (
@@ -27,6 +28,6 @@ ReactDOM.render(
         </div>
       );
     }}
-  </CuriProvider>,
+  </Router>,
   document.getElementById("root")
 );