+++ examples/react/modal/src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { curi } from "@curi/router";
 import Browser from "@hickory/browser";
-import { CuriProvider } from "@curi/react-dom";
+import { curiProvider } from "@curi/react-dom";
 
 import routes from "./routes";
 import Display from "./components/Display";
@@ -10,9 +10,10 @@ import NavLinks from "./components/NavLinks";
 
 const history = Browser();
 const router = curi(history, routes);
+const Router = curiProvider(router);
 
 ReactDOM.render(
-  <CuriProvider router={router}>
+  <Router>
     {({ response, navigation }) => (
       <div>
         <NavLinks />
@@ -23,6 +24,6 @@ ReactDOM.render(
         />
       </div>
     )}
-  </CuriProvider>,
+  </Router>,
   document.getElementById("root")
 );