+++ examples/react/transitions/src/index.js
@@ -3,16 +3,17 @@ import ReactDOM from "react-dom";
 import { curi } from "@curi/router";
 import Browser from "@hickory/browser";
 import { TransitionGroup, CSSTransition } from "react-transition-group";
-import { CuriProvider } from "@curi/react-dom";
+import { curiProvider } from "@curi/react-dom";
 
 import routes from "./routes";
 import NavLinks from "./components/NavLinks";
 
 const history = Browser();
 const router = curi(history, routes);
+const Router = curiProvider(router);
 
 ReactDOM.render(
-  <CuriProvider router={router}>
+  <Router>
     {({ response }) => {
       const { location, params, body: Body } = response;
       return (
@@ -30,6 +31,6 @@ ReactDOM.render(
         </div>
       );
     }}
-  </CuriProvider>,
+  </Router>,
   document.getElementById("root")
 );